% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwilenalim-package.R
\docType{package}
\name{gwilenalim-package}
\alias{gwilenalim}
\alias{gwilenalim-package}
\title{gwilenalim: Package Pour Alimenter Geogwilen Depuis Des Bases De Données Externes}
\description{
Ce package permet de rappatrier et de mettre en forme des données d'autres producteurs pour geogwilen.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AnthonyDEBUR/gwilenalim}
  \item Report bugs at \url{https://github.com/AnthonyDEBUR/gwilenalim/issues}
}

}
\author{
\strong{Maintainer}: Anthony De Burghrave \email{anthony.deburghrave@eaux-et-vilaine.bzh}

}
\keyword{internal}
