% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_data_step.R
\name{importe_data_step}
\alias{importe_data_step}
\title{importe_data_step}
\usage{
importe_data_step(dossier)
}
\arguments{
\item{dossier}{: adresse du dossier où sont stockés les fichiers csv exportés depuis le site
https://www.assainissement.developpement-durable.gouv.fr/}
}
\value{
statut de l'enregistrement des données (OK si les lignes ont été importées et KO sinon)
}
\description{
importe_data_step
}
\examples{

  config <- yaml::read_yaml("C://workspace//gwilenalim//yaml//config.yml")
  dossier <- system.file(package = "gwilenalim")
  importe_data_step(dossier)

}
