% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_redevances_industrie.R
\name{importe_redevances_industrie}
\alias{importe_redevances_industrie}
\title{importe_redevances_industrie}
\usage{
importe_redevances_industrie(fichier)
}
\arguments{
\item{fihier}{: adresse du fichier où est stocké les fichiers xlsx dézippé depuis le site AELB}
}
\value{
statut de l'enregistrement des données (OK si les lignes ont été importées et KO sinon)
}
\description{
Fonction pour importer les données redevances industrielles de l'AELB disponibles sous https://donnees-documents.eau-loire-bretagne.fr/home/donnees/flux-de-pollution---industries.html
}
\examples{
config <- yaml::read_yaml("C://workspace//gwilenalim//yaml//config.yml")
 dossier <- system.file(package = "gwilenalim")
fichier0<-paste0(dossier, "/Industrie/TAB_Effluents_industriels_rejetes_11e_programme.xlsx")

 
 
importe_redevances_industrie(fichier0)
}
