% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_stations_qualite.R
\name{importe_stations_qualite}
\alias{importe_stations_qualite}
\title{importe_stations_qualite}
\usage{
importe_stations_qualite()
}
\value{
statut de l'enregistrement des données (OK si les lignes ont été importées et KO sinon)
}
\description{
Fonction pour importer les coordonnées des stations qualité depuis le SANDRE et compléter le référentiel avec le référentiel AELB pour les réseaux d'appartenance. On se limite aux stations sur l'emprise du bassin de la Vilaine.
Réferentiel AELB disponible ici : https://donnees-documents.eau-loire-bretagne.fr/sites/donnees-documents/home/donnees/referentiel-stations-mesure-qualite.html
Référentiel métropolitain des stations ici : https://www.sandre.eaufrance.fr/atlas/srv/fre/catalog.search#/metadata/71767e88-a021-4e88-8787-5feed04958d6
}
\examples{
importe_stations_qualite()
}
