% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sage_vilaine.R
\docType{data}
\name{sage_vilaine}
\alias{sage_vilaine}
\title{sage_vilaine}
\format{
A sf object with 1 row and 2 variables
\describe{
\item{ nom }{  character }
\item{ geom }{  sfc_MULTIPOLYGON }
}
}
\source{
https://sandre.eaufrance.fr/atlas/srv/fre/catalog.search#/metadata/2faf7c9e-ad62-4137-b80b-8b0caa3609cb
}
\usage{
sage_vilaine
}
\description{
data in 'datasets'.
}
\keyword{datasets}
